agentcore-deployment/requirements.txt
# Core dependencies for blockchain data processing agent
strands-agents>=0.1.0
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0

# MCP (Model Context Protocol) support
mcp>=1.0.0

# AWS SDK for additional AWS service integration
boto3>=1.35.0
botocore>=1.35.0

# OpenTelemetry for observability (recommended for AgentCore)
aws-opentelemetry-distro>=0.10.1

# Additional utilities
pydantic>=2.0.0
httpx>=0.25.0